langchain-templates/langchain-output-parsers/pydantic-output-parser/pyproject.toml
[project]
name = "pydantic-output-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "python-dotenv>=1.1.1",
]
